fix(search,tools): remove answer + structured_search tools (#301)
These were the only consumers of the provider-independent AnswerProvider/
StructuredProvider capability interfaces and Exa's Answer/StructuredSearch
methods; #302's research_panel tool needs that interface slot. Exa's Web/
News/Scholarly capabilities and the shared unconfiguredProviderError/callTool
test helpers used by other tools are untouched.
