+++ internal/search/exa.go
@@ -8,35 +8,24 @@ import (
 	"io"
 	"net/http"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/zoharbabin/web-researcher-mcp/internal/circuit"
 )
 
 // Exa (exa.ai) is a neural/semantic search API exposing a single rich /search
-// endpoint plus grounded /answer and decoupled /contents. This file implements
-// every Exa surface the server uses, all sharing one breaker-wrapped, x-api-key
-// authenticated HTTP core:
+// endpoint plus decoupled /contents. This file implements every Exa surface
+// the server uses, all sharing one breaker-wrapped, x-api-key authenticated
+// HTTP core:
 //
 //   - search.Provider           (Web, News, Images)  — web_search / news_search
 //   - search.AcademicProvider    (Scholarly)          — academic_search via category:"research paper"
-//   - search.AnswerProvider      (Answer)             — the provider-independent `answer` tool
-//   - search.StructuredProvider  (StructuredSearch)   — the provider-independent `structured_search` tool
-//
-// The synthesis interfaces are vendor-neutral (synthesis.go); Exa is just the
-// first implementer. A future provider (e.g. Perplexity) implements the same
-// interfaces and the tools pick it up with no tool-layer change.
 //
 // All facts encoded here were live-verified against api.exa.ai (2026-06-04):
 //   - auth is the `x-api-key` header (NOT Authorization: Bearer — docs list both).
 //   - content fields MUST nest under "contents" ({"contents":{"highlights":true}});
 //     top-level text/highlights/summary returns HTTP 400.
-//   - structured extraction is `contents.summary.schema` (a JSON Schema) → each
-//     results[].summary comes back as a JSON STRING matching it (there is no
-//     top-level output/grounding block — that was a stale-doc claim).
 //   - news = category:"news" + startPublishedDate → dated results.
-//   - entities[] is a PER-RESULT field, present only for category:"company".
 const (
 	exaBaseURL = "https://api.exa.ai"
 
@@ -235,144 +224,6 @@ func (e *ExaProvider) doScholarly(ctx context.Context, params AcademicSearchPara
 	return results, nil
 }
 
-// --- Generic synthesis capabilities (AnswerSearcher / StructuredSearcher) ----
-//
-// Exa implements the provider-independent AnswerSearcher and StructuredSearcher
-// interfaces (synthesis.go). The `answer` and `structured_search` tools talk to
-// those interfaces only; Exa is just the first implementer.
-
-// Answer returns a grounded natural-language answer with citations via /answer.
-func (e *ExaProvider) Answer(ctx context.Context, params AnswerParams) (*AnswerResult, error) {
-	var out *AnswerResult
-	err := e.deps.Breaker.Execute(func() error {
-		body := map[string]any{"query": params.Query}
-		var resp exaAnswerResponse
-		if er := e.doRequest(ctx, "/answer", body, &resp); er != nil {
-			return er
-		}
-		res := &AnswerResult{Answer: resp.Answer, Provider: "exa", CostUSD: resp.CostDollars.Total}
-		for _, c := range resp.Citations {
-			res.Citations = append(res.Citations, Citation(c))
-		}
-		out = res
-		return nil
-	})
-	return out, err
-}
-
-// StructuredSearch runs /search with an optional category and an optional
-// summary JSON Schema, surfacing per-result entities[] when present. It
-// validates Exa's own constraints (category vocabulary, summary-schema limits)
-// up front and returns an InvalidParamsError (not retryable, no breaker trip)
-// so a malformed request never burns a paid call on a guaranteed 400.
-func (e *ExaProvider) StructuredSearch(ctx context.Context, params StructuredParams) (*StructuredResult, error) {
-	if msg := e.validateStructured(params); msg != "" {
-		return nil, &InvalidParamsError{Provider: "exa", Message: msg}
-	}
-	var out *StructuredResult
-	err := e.deps.Breaker.Execute(func() error {
-		contents := map[string]any{"highlights": true}
-		if len(params.Schema) > 0 {
-			// A schema turns the summary into structured JSON conforming to it.
-			contents["summary"] = map[string]any{"schema": params.Schema}
-		} else {
-			contents["summary"] = true
-		}
-		body := map[string]any{
-			"query":      params.Query,
-			"type":       exaSearchType,
-			"numResults": clamp(params.NumResults, 1, exaMaxResults),
-			"contents":   contents,
-		}
-		if params.Category != "" {
-			body["category"] = params.Category
-		}
-
-		var resp exaSearchResponse
-		if er := e.doRequest(ctx, "/search", body, &resp); er != nil {
-			return er
-		}
-		res := &StructuredResult{Provider: "exa", CostUSD: resp.CostDollars.Total}
-		for _, r := range resp.Results {
-			item := StructuredItem{
-				Title:         r.Title,
-				URL:           r.URL,
-				PublishedDate: r.PublishedDate,
-				Author:        r.Author,
-				Highlights:    r.Highlights,
-				Summary:       jsonOrString(r.Summary),
-				Entities:      r.Entities,
-			}
-			res.Results = append(res.Results, item)
-		}
-		out = res
-		return nil
-	})
-	return out, err
-}
-
-// ExaCategories returns the structured-search categories Exa supports (live-
-// verified to return HTTP 200). Empty string (no category) is always allowed and
-// is not listed. Exposed so a caller can surface valid options without
-// hard-coding Exa's vocabulary.
-func ExaCategories() []string {
-	return []string{"company", "people", "research paper", "news", "pdf", "github", "financial report", "personal site"}
-}
-
-// validateStructured enforces Exa's own request constraints before any paid
-// call: a known category (or none), and a summary schema within Exa's documented
-// limits (root object, ≤10 properties, nesting depth ≤2, primitive array items).
-// Returns "" when valid, else a human-readable reason. This vendor-specific
-// knowledge lives here, in the provider — the generic tool stays vendor-neutral.
-func (e *ExaProvider) validateStructured(params StructuredParams) string {
-	if params.Category != "" {
-		ok := false
-		for _, c := range ExaCategories() {
-			if c == params.Category {
-				ok = true
-				break
-			}
-		}
-		if !ok {
-			return fmt.Sprintf("invalid category %q for provider exa; allowed: %s (or omit)",
-				params.Category, strings.Join(ExaCategories(), ", "))
-		}
-	}
-	if len(params.Schema) == 0 {
-		return ""
-	}
-	var schema map[string]any
-	if err := json.Unmarshal(params.Schema, &schema); err != nil {
-		return "schema is not valid JSON: " + err.Error()
-	}
-	if t, _ := schema["type"].(string); t != "object" {
-		return `schema root must have "type": "object"`
-	}
-	props, ok := schema["properties"].(map[string]any)
-	if !ok || len(props) == 0 {
-		return `schema must declare a non-empty "properties" object`
-	}
-	if len(props) > 10 {
-		return fmt.Sprintf("schema has %d properties; exa allows at most 10", len(props))
-	}
-	for name, p := range props {
-		pm, ok := p.(map[string]any)
-		if !ok {
-			return fmt.Sprintf("property %q must be a schema object", name)
-		}
-		switch pm["type"] {
-		case "object":
-			return fmt.Sprintf("property %q is a nested object; exa schemas allow nesting depth at most 2 (use a flat object)", name)
-		case "array":
-			items, _ := pm["items"].(map[string]any)
-			if it, _ := items["type"].(string); it == "object" || it == "array" {
-				return fmt.Sprintf("property %q is an array of %s; exa requires array items to be primitive", name, it)
-			}
-		}
-	}
-	return ""
-}
-
 // --- Shared HTTP core -------------------------------------------------------
 
 // doRequest POSTs a JSON payload to an Exa endpoint with x-api-key auth and
@@ -422,19 +273,17 @@ func (e *ExaProvider) doRequest(ctx context.Context, path string, payload map[st
 // --- Response shapes (live-verified field names) ----------------------------
 
 type exaSearchResponse struct {
-	Results     []exaResult    `json:"results"`
-	CostDollars exaCostDollars `json:"costDollars"`
+	Results []exaResult `json:"results"`
 }
 
 type exaResult struct {
-	Title         string          `json:"title"` // may be ""
-	URL           string          `json:"url"`
-	PublishedDate string          `json:"publishedDate"` // may be ""/null
-	Author        string          `json:"author"`        // may be ""/null
-	Highlights    []string        `json:"highlights"`
-	Summary       string          `json:"summary"`  // a JSON string when a summary schema was supplied
-	Entities      json.RawMessage `json:"entities"` // present only for category:"company"
-	Score         float64         `json:"score"`    // relevance/quality score 0-1
+	Title         string   `json:"title"` // may be ""
+	URL           string   `json:"url"`
+	PublishedDate string   `json:"publishedDate"` // may be ""/null
+	Author        string   `json:"author"`        // may be ""/null
+	Highlights    []string `json:"highlights"`
+	Summary       string   `json:"summary"` // a JSON string when a summary schema was supplied
+	Score         float64  `json:"score"`   // relevance/quality score 0-1
 }
 
 // snippet returns the best available short text for a result: the first
@@ -444,24 +293,6 @@ func (r exaResult) snippet() string {
 		return r.Highlights[0]
 	}
 	return r.Summary
-}
-
-type exaAnswerResponse struct {
-	Answer      string         `json:"answer"`
-	Citations   []exaCitation  `json:"citations"`
-	CostDollars exaCostDollars `json:"costDollars"`
-}
-
-type exaCitation struct {
-	Title         string `json:"title"`
-	URL           string `json:"url"`
-	PublishedDate string `json:"publishedDate"`
-}
-
-// exaCostDollars captures Exa's per-call cost estimate (total in USD). It is an
-// estimate, not an invoice, and is surfaced into audit metadata as cost_usd.
-type exaCostDollars struct {
-	Total float64 `json:"total"`
 }
 
 // --- Helpers ----------------------------------------------------------------
@@ -506,30 +337,7 @@ func publishYear(iso string) int {
 	return 0
 }
 
-// jsonOrString normalizes Exa's summary string for embedding. When a summary
-// schema was supplied, Exa returns a JSON object/array string — embedded
-// verbatim so the caller gets structured data. Otherwise (plain-text summary,
-// or a bare JSON scalar like "null"/"123" that only coincidentally parses) it is
-// encoded as a JSON string, preserving the "object-or-string" output contract.
-// Empty → nil (dropped by omitempty).
-func jsonOrString(summary string) json.RawMessage {
-	if summary == "" {
-		return nil
-	}
-	if trimmed := strings.TrimSpace(summary); len(trimmed) > 0 &&
-		(trimmed[0] == '{' || trimmed[0] == '[') && json.Valid([]byte(summary)) {
-		return json.RawMessage(summary)
-	}
-	encoded, err := json.Marshal(summary)
-	if err != nil {
-		return nil
-	}
-	return encoded
-}
-
 var (
-	_ Provider           = (*ExaProvider)(nil)
-	_ AcademicProvider   = (*ExaProvider)(nil)
-	_ AnswerProvider     = (*ExaProvider)(nil)
-	_ StructuredProvider = (*ExaProvider)(nil)
+	_ Provider         = (*ExaProvider)(nil)
+	_ AcademicProvider = (*ExaProvider)(nil)
 )
