+++ internal/tools/metadata_test.go
@@ -38,8 +38,6 @@ var expectedTools = []string{
 	"awesome_list_search",
 	"local_search",
 	"monarch_search",
-	"answer",
-	"structured_search",
 	"get_my_analytics",
 	"memory_save",
 	"memory_recall",
@@ -272,8 +270,6 @@ func TestOutputSchemaMatchesResponse(t *testing.T) {
 		"patent_search":       {"query": "test"},
 		"sequential_search":   {"searchStep": "initial research", "stepNumber": 1, "nextStepNeeded": false},
 		"citation_graph":      {"paper": "10.1/x"},
-		"answer":              {"query": "test"},
-		"structured_search":   {"query": "test"},
 		"format_bibliography": {"sources": []any{map[string]any{"url": "https://example.com/a", "title": "A", "author": "Smith, J.", "date": "2024"}}},
 		"audit_bibliography":  {"entries": []any{map[string]any{"url": "https://example.com/a", "title": "A", "doi": "10.1/x"}}},
 		// setupTestDeps has a nil LinkVerifier → archive_source returns status:"unavailable",
@@ -370,8 +366,6 @@ func TestExternalContentToolsCarryTrustMarker(t *testing.T) {
 		"citation_graph":      "untrusted-external-content",
 		"verify_citation":     "untrusted-external-content",
 		"audit_bibliography":  "untrusted-external-content",
-		"answer":              "untrusted-external-content",
-		"structured_search":   "untrusted-external-content",
 		"filing_search":       "untrusted-external-content",
 		"legal_search":        "untrusted-external-content",
 		"econ_search":         "untrusted-external-content",
@@ -395,8 +389,6 @@ func TestExternalContentToolsCarryTrustMarker(t *testing.T) {
 		"citation_graph":      {"paper": "10.1/x"},
 		"verify_citation":     {"citation": "https://example.com/paper"},
 		"audit_bibliography":  {"entries": []any{map[string]any{"url": "https://example.com/paper", "title": "A"}}},
-		"answer":              {"query": "test"},
-		"structured_search":   {"query": "test"},
 		"filing_search":       {"query": "AAPL"},
 		"legal_search":        {"query": "miranda"},
 		"econ_search":         {"series_id": "GDP"},
