+++ internal/tools/tools_test.go
@@ -130,36 +130,6 @@ func newTestBreaker() *circuit.Breaker {
 	return circuit.New(circuit.Config{FailureThreshold: 5, ResetTimeout: 60})
 }
 
-// mockSynthProvider implements the provider-independent AnswerProvider and
-// StructuredProvider interfaces, so wiring it into AnswerProviders/
-// StructuredProviders makes the conditionally-registered `answer` and
-// `structured_search` tools visible to the CI drift tests. It is vendor-neutral
-// (Name "mocksynth") — the tools must work for any provider, not just Exa.
-type mockSynthProvider struct{}
-
-func (m *mockSynthProvider) Name() string { return "mocksynth" }
-
-func (m *mockSynthProvider) Metadata() search.ProviderMeta {
-	return search.ProviderMeta{Regions: []string{"*"}, RateClass: "free", Description: "mock synthesis provider"}
-}
-
-func (m *mockSynthProvider) Answer(_ context.Context, _ search.AnswerParams) (*search.AnswerResult, error) {
-	return &search.AnswerResult{
-		Answer:    "A test answer.",
-		Citations: []search.Citation{{Title: "Source", URL: "https://example.com", PublishedDate: "2026-01-01"}},
-		Provider:  "mocksynth",
-		CostUSD:   0.005,
-	}, nil
-}
-
-func (m *mockSynthProvider) StructuredSearch(_ context.Context, p search.StructuredParams) (*search.StructuredResult, error) {
-	return &search.StructuredResult{
-		Results:  []search.StructuredItem{{Title: "Item", URL: "https://example.com"}},
-		Provider: "mocksynth",
-		CostUSD:  0.007,
-	}, nil
-}
-
 // mockAcademicProvider implements AcademicProvider + CitationSearcher, so wiring
 // it into AcademicProviders makes citation_graph register and exercises the
 // academic_search academic path in tests.
@@ -340,7 +310,6 @@ func (m *mockArchiveResolver) Lookup(_ context.Context, domain string, _ int) ([
 }
 
 func setupTestDeps() Dependencies {
-	synth := &mockSynthProvider{}
 	academic := &mockAcademicProvider{}
 	filing := &mockFilingProvider{}
 	caseProv := &mockCaseProvider{}
@@ -352,8 +321,6 @@ func setupTestDeps() Dependencies {
 	return Dependencies{
 		Cache:                cache.NewNoop(),
 		Search:               &mockProvider{},
-		AnswerProviders:      map[string]search.AnswerProvider{synth.Name(): synth},
-		StructuredProviders:  map[string]search.StructuredProvider{synth.Name(): synth},
 		AcademicProviders:    map[string]search.AcademicProvider{academic.Name(): academic},
 		FilingProviders:      map[string]search.FilingProvider{filing.Name(): filing},
 		CaseProviders:        map[string]search.CaseProvider{caseProv.Name(): caseProv},
