+++ llms-install.md
@@ -113,7 +113,7 @@ The differentiator is the **trust suite**: `verify_citation` checks whether one 
 
 Alongside those, the always-on core tools include web search, full-page/document scraping (`mode: raw` for verbatim source), combined search-and-scrape with quality ranking, image and news search, academic search (real DOIs), patent search (US/EP/WO/JP/CN/KR), US case-law search (`legal_search`, CourtListener), economic data (`econ_search`, FRED + World Bank + OECD + Eurostat), clinical-trial search (`clinical_search`, ClinicalTrials.gov), community-curated GitHub list search (`awesome_list_search`, ecosyste.ms), research-session export and bibliography formatting (APA/MLA/BibTeX/RIS/CSL-JSON), multi-step `sequential_search` with recoverable sessions, and `brand_research` for structured brand identity data (colors, logos, typography, social handles) from any domain or company name.
 
-A few tools register only when their provider or config is present: `citation_graph` requires a citation-capable academic provider (OpenAlex or Semantic Scholar); `filing_search` (SEC EDGAR) requires `EDGAR_CONTACT_EMAIL` (falls back to `OPENALEX_EMAIL`); `answer` and `structured_search` require `EXA_API_KEY`; `local_search` (place/business search) registers only when `BRAVE_API_KEY` is set. Optional enrichment keys `COURTLISTENER_API_TOKEN` and `FRED_API_KEY` raise limits or add data to always-available tools.
+A few tools register only when their provider or config is present: `citation_graph` requires a citation-capable academic provider (OpenAlex or Semantic Scholar); `filing_search` (SEC EDGAR) requires `EDGAR_CONTACT_EMAIL` (falls back to `OPENALEX_EMAIL`); `local_search` (place/business search) registers only when `BRAVE_API_KEY` is set. Optional enrichment keys `COURTLISTENER_API_TOKEN` and `FRED_API_KEY` raise limits or add data to always-available tools.
 
 Operators can additionally enable opt-in, consent-gated tools (per-user analytics, long-term memory, shared workspaces) that register only when their feature is turned on.
