+++ python/web_researcher_mcp/__init__.py
@@ -25,8 +25,6 @@ from web_researcher_mcp.models import (  # noqa: F401
     AcademicSearchPaper,
     AcademicSearchResponse,
     Analytics,
-    AnswerCitation,
-    AnswerResponse,
     ArchiveResult,
     ArchiveSourceResponse,
     AuditBibliographyEntry,
@@ -120,8 +118,6 @@ from web_researcher_mcp.models import (  # noqa: F401
     SizeMetadata,
     Social,
     StructuredData,
-    StructuredSearchResponse,
-    StructuredSearchResult,
     Summary,
     SyllabusSearchResponse,
     SyllabusSearchResult,
@@ -150,8 +146,6 @@ __all__ = [
     "AcademicSearchPaper",
     "AcademicSearchResponse",
     "Analytics",
-    "AnswerCitation",
-    "AnswerResponse",
     "ArchiveResult",
     "ArchiveSourceResponse",
     "AuditBibliographyEntry",
@@ -245,8 +239,6 @@ __all__ = [
     "SizeMetadata",
     "Social",
     "StructuredData",
-    "StructuredSearchResponse",
-    "StructuredSearchResult",
     "Summary",
     "SyllabusSearchResponse",
     "SyllabusSearchResult",
